BOT/requirements.txt
aiogram>=3.20.0.post0
APScheduler>=3.11.0
python-dotenv>=1.1.0
SQLAlchemy>=2.0.41
pandas>=2.2.3
openpyxl>=3.1.5